Summary: Adding widgets and slot code for the V1718 pulsers
Details:

New tab to provide full control over the two V1718 digital pulses that
are provided onboard with the V1718. The widget code is done and
signals/slots mostly implemented. Next step is to complete the ADAQ
libraries to support a V1718 "bridge manager" to handle the interface
commands to the hardare.