Summary: Starting new refactoring branch for major code overhaul
Details:

:: Sigh :: This day was inevitable. Significant new features for
ADAQAcquisition are required so it's as good a time as any to give the
code a complete refactoring overhaul. The purpose of refactoring is
(at present) to be:

-Major modularization of code into many classes
-Managers to handle graphics, acquisition
-Code standardization (with ADAQAnalysis)
-Eliminate hardcoding, make flexible
-Update/improve code documentation

I have started with this checkin by splitting ADAQAcquisition.cc/.hh
into main C++ ADAQAcquisition.cc and AAInterface.cc/hh, as well as
adding all of the class structures planned for us. The makefile has
been completely updated and a few header files have moved around.