Starting branch to integrate the new ADAQ readout library; prototype integration complete
Details:

The ADAQ readout library (libADAQReadout.so) is an extension of the
ADAQ libraries. It provides flexible yet standardized readout
functionality for data acquisition with CAEN hardware for persistent
storage and offline analysis. It is principally designed to serve as a
power conduit between ADAQAcquisition and ADAQAnalysis but is made
general enough for any other readout code.

Initial prototype integrating these methods into AAAcquisitionManager
class are complete (minus filling out the ADAQReadoutInformation
class). Next phase needs to involve testing/development on running
hardware to determine if this method actually works and debug.