Adding new persistency subtab widgets to support ADAQReadout options
Details:

ADAQReadout supports readout of raw waveforms, energy data (pulse
height/area), and/or PSD data (total/tail integrals); thus, the user
needs to be able to select which data should be stored in the file.
These widgets are now implemented along with a necessary rearrage of
the group frame containing them.