Essentially finishing ADAQReadout implementation (minus the PSD)
Details:

We are moving rapidly towards finishing the integration of the
ADAQReadout library. Nearly everythign is in place and has been tested
with waveform readout (just baselines ... not actually hooked up to
detector yet). The only thing that is not implemented is readout of
the PSD integrals, but this will have to wait until that functionality
is implemented into AAInterface and AAComputation.

Other changes include:

- Updating ADAQ file name setting algorithms for stricter and more
  correct setting of the file extension

- Changing all ROOT ClassDef() numbers from '0' to '1' to be in
  compliance with ROOT recommendations