Move TGraph new/delete, TGraph settings out of AAGraphics::PlotWaveforms() for efficiency
Details:

The AAGraphics::PlotWaveforms() method is an extremely CPU/memory
intensive portion of the acquisition loop, called as many times as the
system triggers when the user selects to view the waveform (i.e. it's
an oscilloscope). Previously, we deleted/recreated a vector of TGraphs
via new and delete dynamic memory allocation _every time_ this
function was called. We also, as a consequence, had to reset "static"
graphical settings (axis labels, size, colors, etc) every time.
This is extremely inefficient!

Now, a vector of TGraphs is deleted/recreated _once_ at the beginning
of the acquisition loop and all static graphical objects are
set. TGraphs are then drawn via the TGraph::DrawGraph() method. This
enables dynamic changes in number of channels, different hardware, etc
between runs, handles new/delete without memory leaks, and should be
VASTLY more efficient!

Note that this is UNTESTED as of this commit.