Updating for new 'single TTree' format of ADAQReadout; efficiency improvements to acquisition loop
Details:

Main purpose of this commit is to update the AAAcquisitionManager to
be in compliance with the new 'single TTree' data storage format for
ADAQ files. Now, for each digitizer channel, two branches are created
on the same tree: one for the channel waveform, one for the channel
waveform data.

Increased acquisition loop efficiency by moving PSD integral
calculation INSIDE of ultra-rate mode protection. Another important
change is ensuring the trigger time stamp, channel ID, and board ID of
the waveform is stored in the ADAQ file ALWAYS regardless of what
information is desired to be stored. Other small updates to the
acquisition loop were made, including commenting where needed