Updating Makefile to be in compliance with ADAQ libraries version 1.0.0-beta
Details:

This master branch is about to get a massive merge that will
incorporate several massive changes (utilization of new ADAQReadout
library, on-the-fly PSD, new ROOT object output, methods, etc). Before
this happens, I want to update the master branch to be in compliance
with the ADAQ libaries v1.0.0-beta so that the state of the code here
is usable and can be tagged as 0.9.9 before we merge and then tag the
beta for the first official production version.