BUG FIX: Correctly write/close ADAQ file when using acquisition timer
Details:

The write/close for the ADAQ file when using the acquisition timer was
using the old boolean value 'ADAQFileIsOpen', which was a class member
of AAAcquisitionManager. This entire file open check method has moved
to the ADAQ readout manager under the new scheme. I have updated this
functionality, removed the old 'ADAQFileIsOpen' boolean from
AAAcquisitionManager, and verified functionality successfully on
hardware.