Enormous restructuring of acquisition loop to support DPP-PSD readout
Details:

Major structural changes in the acquisition loop contained in the
AAAcquisitionManager::StartAcquisition() method. In short, the
previous readout scheme was

  for(events)
     for(channels)
        ...
	for(samples)
	   ...
	   Get waveform data
	   Analysis full waveforms
	   ...
        Fill spectra, PSD histogram
      Fill ROOT Tree
      Draw waveforms ...
  Plot

This made sense when a single event in standard CAEN firmware always
contained all enabled channels. In DPP-PSD, however, channel
self-triggering is independent so we can have different number of
events on each channel. This require restructuring the order of the
aquisition loop to ensure that we correctly captured all events on all
enabled channels. Now, channels are looped over first following by
events:

  for(channels)
     for(events)
        ...
	etc, etc

The key change is ensure that the ROOT TTree and other items are
filled at the event-level.

AAVMEManager is temporarily hardcoded with DPP-PSD parameters.

The state of the code at this point has readout a full waveform!