Minor updates to get PSD gate / trigger settings to work for channel 0
Details:

Using the pretrigger, gate offset, and long/short gates now works for
the channels. However, there appears to be a bug, possible in the CAEN
firmware, where settings the last channel's PSD pretrigger sets ALL
channel pretriggers. This issue should be resolved with CAEN; perhaps
individual channel settings are not allowed on the DT5790?